# src/modules/chat/chat.model.ts
import { Schema, model } from "mongoose";

const ChatSchema = new Schema(
  {
    senderId: { type: String, required: true },
    receiverId: { type: String, required: true },
    message: { type: String, required: true },
  },
  { timestamps: true }
);

export const ChatMessage = model("ChatMessage", ChatSchema);